RS-15 feat(statistics): add filter new word and rebase getUserWords
